Prep for 1.6.4.
Started by clearing out some old stuff, but it turned out most of it was
fine. The major thing left to do is updating references to methods that
MCP probably renamed, and possibly finding better Forge stuff to help
out.

Note the license change -- I think the MIT license is more open than
MMPL, and even if it weren't, it's a more standard thing to use. The
sounds and textures haven't changed license; this only applies to the
code.

Also note we're dropping the IC2 dependency. Not entirely sure what I
want to do with the power requirements -- instinct says just keeping the
numbers exactly as they were is going to be reasonable, but it needs to
be re-examined. I need to find how I derived the base values in the
first place, though, and possibly make the config friendlier by
providing those instead. We'll see.